Sync NuttX and Apps to branch master (Preparing to clone nuttx to version master... De-initializing submodule nuttx... Removing existing work tree directory: nuttx Cloning nuttx with branch 'master' from https://github.com/apache/nuttx.git... Cloning of nuttx for branch 'master' successful. nuttx commit hash: 8c32bed29e5ac6c0924f400878f8c47035b34107 8c32bed29e5ac6c0924f400878f8c47035b34107, Preparing to clone apps to version master... De-initializing submodule apps... Removing existing work tree directory: apps Cloning apps with branch 'master' from https://github.com/apache/nuttx-apps.git... Cloning of apps for branch 'master' successful. apps commit hash: c19c2a71f3773b19d7ad91ceba6d8d748d12be5f c19c2a71f3773b19d7ad91ceba6d8d748d12be5f)

+++ nuttx/tools/refresh.sh
@@ -1,43 +1,32 @@
 #!/usr/bin/env bash
 # tools/refresh.sh
 #
-#   Copyright (C) 2014, 2016-2017, 2019 Gregory Nutt. All rights reserved.
-#   Author: Gregory Nutt <gnutt@nuttx.org>
+# SPDX-License-Identifier: Apache-2.0
 #
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions
-# are met:
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.  The
+# ASF licenses this file to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance with the
+# License.  You may obtain a copy of the License at
 #
-# 1. Redistributions of source code must retain the above copyright
-#    notice, this list of conditions and the following disclaimer.
-# 2. Redistributions in binary form must reproduce the above copyright
-#    notice, this list of conditions and the following disclaimer in
-#    the documentation and/or other materials provided with the
-#    distribution.
-# 3. Neither the name NuttX nor the names of its contributors may be
-#    used to endorse or promote products derived from this software
-#    without specific prior written permission.
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
-# OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
-# AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
-# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+# License for the specific language governing permissions and limitations
+# under the License.
 #
 
 WD=`test -d ${0%/*} && cd ${0%/*}; pwd`
+CWD=`pwd`
 
 USAGE="USAGE: $0 [options] <board>:<config>+"
 ADVICE="Try '$0 --help' for more information"
 
 unset CONFIGS
+diff=0
 debug=n
 defaults=n
 prompt=y
@@ -82,10 +71,18 @@ while [ ! -z "$1" ]; do
     echo "  --help"
     echo "     Show this help message and exit"
     echo "  <board>"
-    echo "     The board directory under nuttx/boards"
+    echo "     The board directory under nuttx/boards/arch/chip/"
     echo "  <config>"
-    echo "     The board configuration directory under nuttx/boards/<board>/configs"
-    echo "  Note: all configuration is refreshed if <board>:<config> equals all."
+    echo "     The board configuration directory under nuttx/boards/arch/chip/<board>/configs"
+    echo "  <archname>"
+    echo "     The architecture directory under nuttx/boards/"
+    echo "  <chipname>"
+    echo "     The chip family directory under nuttx/boards/<arch>/"
+    echo " "
+    echo "  Note1: all configurations are refreshed if <board>:<config> is replaced with \"all\" keyword" 
+    echo "  Note2: all configurations of arch XYZ are refreshed if \"arch:<namearch>\" is passed"
+    echo "  Note3: all configurations of chip XYZ are refreshed if \"chip:<chipname>\" is passed"
+    echo "  Note4: all configurations of board XYZ are refreshed if \"board:<boardname>\" is passed"
     exit 0
     ;;
   * )
@@ -110,36 +107,6 @@ if [ ! -x tools/${MYNAME} ] ; then
   exit 1
 fi
 
-# If the cmpconfig executable does not exist, then build it
-
-CMPCONFIG_TARGET=cmpconfig
-CMPCONFIG1=tools/cmpconfig
-CMPCONFIG2=tools/cmpconfig.exe
-CMPCONFIGMAKEFILE=Makefile.host
-CMPCONFIGMAKEDIR=tools
-
-if [ -x ${CMPCONFIG1} ]; then
-  CMPCONFIG=${CMPCONFIG1}
-else
-  if [ -x ${CMPCONFIG2} ]; then
-    CMPCONFIG=${CMPCONFIG2}
-  else
-    make -C ${CMPCONFIGMAKEDIR} -f ${CMPCONFIGMAKEFILE} ${CMPCONFIG_TARGET} || \
-      { echo "ERROR: make ${CMPCONFIG1} failed" ; exit 1 ; }
-  fi
-fi
-
-if [ -x ${CMPCONFIG1} ]; then
-  CMPCONFIG=${CMPCONFIG1}
-else
-  if [ -x ${CMPCONFIG2} ]; then
-    CMPCONFIG=${CMPCONFIG2}
-  else
-    echo "ERROR: Failed to create  ${CMPCONFIG1}"
-    exit 1
-  fi
-fi
-
 # Get the board configuration
 
 if [ -z "${CONFIGS}" ]; then
@@ -150,36 +117,73 @@ if [ -z "${CONFIGS}" ]; then
 fi
 
 if [ "X${CONFIGS}" == "Xall" ]; then
-  CONFIGS=`find boards -name defconfig | cut -d'/' -f4,6`
+  echo "Normalizing all boards!"
+  CONFIGS=(`find boards -name defconfig | cut -d'/' -f4,6`)
+else
+  if [[ "X${CONFIGS}" == "Xarch:"* ]]; then
+    IFS=: read -r atype archname <<< "${CONFIGS}"
+    ARCH=$archname
+    echo "Normalizing all boards in arch: ${ARCH} !"
+    CONFIGS=(`find boards/${ARCH} -name defconfig | cut -d'/' -f4,6`)
+  else
+    if [[ "X${CONFIGS}" == "Xchip:"* ]]; then
+      IFS=: read -r atype chipname <<< "${CONFIGS}"
+      CHIP=$chipname
+      echo "Normalizing all boards in chip: ${CHIP} !"
+      CONFIGS=(`find boards/*/${CHIP} -name defconfig | cut -d'/' -f4,6`)
+    else
+      if [[ "X${CONFIGS}" == "Xboard:"* ]]; then
+        IFS=: read -r atype boardname <<< "${CONFIGS}"
+        BOARD=$boardname
+        echo "Normalizing all configs in board: ${BOARD} !"
+	# ATTENTION: It assumes we don't have a board with same name in other arch or chip
+        CONFIGS=(`find boards/*/*/${boardname} -name defconfig | cut -d'/' -f4,6`)
+      fi
+    fi
+  fi
 fi
 
-for CONFIG in ${CONFIGS}; do
-  echo "Refresh ${CONFIG}"
+for i in ${!CONFIGS[@]}; do
+  echo "  [$((${i} + 1))/${#CONFIGS[@]}] Normalize ${CONFIGS[$i]}"
 
   # Set up the environment
 
-  CONFIGSUBDIR=`echo ${CONFIG} | cut -s -d':' -f2`
+  CONFIGSUBDIR=`echo ${CONFIGS[$i]} | cut -s -d':' -f2`
   if [ -z "${CONFIGSUBDIR}" ]; then
-    CONFIGSUBDIR=`echo ${CONFIG} | cut -s -d'/' -f2`
+    CONFIGSUBDIR=`echo ${CONFIGS[$i]} | cut -s -d'/' -f2`
     if [ -z "${CONFIGSUBDIR}" ]; then
-      echo "ERROR: Malformed configuration: ${CONFIG}"
+      echo "ERROR: Malformed configuration: ${CONFIGS[$i]}"
       echo $USAGE
       echo $ADVICE
       exit 1
     else
-      BOARDSUBDIR=`echo ${CONFIG} | cut -d'/' -f1`
+      BOARDSUBDIR=`echo ${CONFIGS[$i]} | cut -d'/' -f1`
     fi
   else
-    BOARDSUBDIR=`echo ${CONFIG} | cut -d':' -f1`
+    BOARDSUBDIR=`echo ${CONFIGS[$i]} | cut -d':' -f1`
   fi
 
-  BOARDDIR=boards/*/*/$BOARDSUBDIR
+  BOARDDIR=${CONFIGS[$i]}
+  if [ ! -d $BOARDDIR ]; then
+    BOARDDIR="${CWD}/${BOARDDIR}"
+  fi
+
+  if [ -d $BOARDDIR ]; then
+    CONFIGSUBDIR=`basename ${CONFIGS[$i]}`
+    BOARDDIR=$(dirname `dirname ${BOARDDIR}`)
+  else
+    BOARDDIR=boards/*/*/$BOARDSUBDIR
+  fi
+
   SCRIPTSDIR=$BOARDDIR/scripts
   MAKEDEFS1=$SCRIPTSDIR/Make.defs
 
   CONFIGDIR=$BOARDDIR/configs/$CONFIGSUBDIR
   DEFCONFIG=$CONFIGDIR/defconfig
   MAKEDEFS2=$CONFIGDIR/Make.defs
+
+  SCRIPTSDIR2=$BOARDDIR/../common/scripts
+  MAKEDEFS3=$SCRIPTSDIR2/Make.defs
 
   # Check the board configuration directory
 
@@ -204,9 +208,19 @@ for CONFIG in ${CONFIGS}; do
     if [ -r $MAKEDEFS1 ]; then
       MAKEDEFS=$MAKEDEFS1
     else
-      echo "No readable Make.defs file at $MAKEDEFS1 or $MAKEDEFS2"
-      exit 1
+      if [ -r $MAKEDEFS3 ]; then
+        MAKEDEFS=$MAKEDEFS3
+      else
+        echo "No readable Make.defs file at $MAKEDEFS1 or $MAKEDEFS2 or $MAKEDEFS3"
+        exit 1
+      fi
     fi
+  fi
+
+  # skip refresh if defconfig contains `#include`
+  if grep -q "#include" $DEFCONFIG; then
+    echo "Note: skipping refresh for debug defconfig."
+    continue
   fi
 
   # Copy the .config and Make.defs to the toplevel directory
@@ -234,61 +248,71 @@ for CONFIG in ${CONFIGS}; do
     # Then run oldconfig or oldefconfig
 
     if [ "X${defaults}" == "Xy" ]; then
-      if [ "X${debug}" = "Xy" ]; then
+      if [ "X${debug}" == "Xy" ]; then
         make olddefconfig V=1
       else
         make olddefconfig 1>/dev/null
       fi
     else
-      if [ "X${debug}" = "Xy" ]; then
+      if [ "X${debug}" == "Xy" ]; then
         make oldconfig V=1
       else
-        make oldconfig 1>/dev/null
+        make oldconfig
       fi
     fi
   fi
 
   # Run savedefconfig to create the new defconfig file
 
-  if [ "X${debug}" = "Xy" ]; then
+  if [ "X${debug}" == "Xy" ]; then
     make savedefconfig V=1
   else
     make savedefconfig 1>/dev/null
   fi
 
-  # Save the refreshed configuration
+  # Show differences
 
-  if [ "X${prompt}" == "Xy" ]; then
+  if ! diff $DEFCONFIG defconfig; then
 
-    # Show differences
+    # Save the refreshed configuration
 
-    $CMPCONFIG $DEFCONFIG defconfig
+    if [ "X${prompt}" == "Xy" ]; then
 
-    read -p "Save the new configuration (y/n)?" -n 1 -r
-    echo
-    if [[ $REPLY =~ ^[Yy]$ ]]
-    then
+      read -p "Save the new configuration (y/n)?" -n 1 -r
+      echo
+      if [[ $REPLY =~ ^[Yy]$ ]]; then
+        echo "Saving the new configuration file"
+        mv defconfig $DEFCONFIG || \
+            { echo "ERROR: Failed to move defconfig to $DEFCONFIG"; exit 1; }
+        chmod 644 $DEFCONFIG
+      fi
+    else
       echo "Saving the new configuration file"
       mv defconfig $DEFCONFIG || \
           { echo "ERROR: Failed to move defconfig to $DEFCONFIG"; exit 1; }
       chmod 644 $DEFCONFIG
     fi
-  else
-    echo "Saving the new configuration file"
-    mv defconfig $DEFCONFIG || \
-        { echo "ERROR: Failed to move defconfig to $DEFCONFIG"; exit 1; }
-    chmod 644 $DEFCONFIG
+
+    diff=1
   fi
 
   # Restore any previous .config and Make.defs files
-
-  if [ -e SAVEconfig ]; then
-    mv SAVEconfig .config || \
-      { echo "ERROR: Failed to move SAVEconfig to .config"; exit 1; }
-  fi
 
   if [ -e SAVEMake.defs ]; then
     mv SAVEMake.defs Make.defs || \
       { echo "ERROR: Failed to move SAVEMake.defs to Make.defs"; exit 1; }
   fi
+
+  if [ -e SAVEconfig ]; then
+    mv SAVEconfig .config || \
+      { echo "ERROR: Failed to move SAVEconfig to .config"; exit 1; }
+
+    if [ "X${debug}" == "Xy" ]; then
+      ./tools/sethost.sh V=1
+    else
+      ./tools/sethost.sh 1>/dev/null
+    fi
+  fi
 done
+
+exit $diff
